Keep pages/404.js to be able to dynamically render it anyway (#89263)
Needed for https://github.com/vercel/vercel/pull/14795

Keep the file so that we can instead dynamically render it as opposed to having to bundle the prerender into the function

+++ packages/next/src/build/index.ts
@@ -3179,6 +3179,11 @@ export default async function build(
 
           // remove server bundles that were exported
           for (const page of staticPages) {
+            if (page === '/404') {
+              // keep the pages 404 chunk as we might need it if we do want to runtime-render the
+              // 404 in the function later
+              continue
+            }
             const serverBundle = getPagePath(page, distDir, undefined, false)
             await fs.unlink(serverBundle)
           }
